// src/app/models/attendance/attreqdocument.ts
export class ATTReqdocumentModel {
    constructor() {

    }
    company_code: string = "";
    document_id: number = 0;
    job_id: string = "";
    job_type: string = "";
    document_name: string = "";
    document_type: string = "";
    document_path: string = "";
    created_by: string = "";
    created_date: string = "";
}